Refactor all networks
